requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
boto3>=1.28.64
prometheus-fastapi-instrumentator>=6.1.0
pytest>=7.4.2
httpx>=0.25.0
uuid>=1.30
aiofiles>=23.2.1
loguru>=0.7.0